+++ storage_drivers/ontap/ontap_san_economy.go
@@ -1424,7 +1424,7 @@ func (d *SANEconomyStorageDriver) mapOntapSANLUN(volConfig *storage.VolumeConfig
 		return err
 	}
 
-	err = PopulateOntapLunMapping(d.API, &d.Config, d.ips, volConfig, lunID)
+	err = PopulateOntapLunMapping(d.API, &d.Config, d.ips, volConfig, lunID, lunPath, d.Config.IgroupName)
 	if err != nil {
 		return fmt.Errorf("error mapping LUN for %s driver: %v", d.Name(), err)
 	}