+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 **Fixes:**
 - **Kubernetes** Change Trident node server to use downward API instead of relying on Kubernetes DNS to find Trident service. (Issue [#328](https://github.com/NetApp/trident/issues/328))
 - Fixes FlexGroup volume deletion against ONTAP version 9.7 (Issue [#326](https://github.com/NetApp/trident/issues/326))
+- Fixed SLM support for ONTAP iSCSI drivers. (Issue [#321](https://github.com/NetApp/trident/issues/321))
 
 ## v20.01.0